Update Zenodo instructions in release checklist (#12)
After an update to our Authorship Guidelines, "Fatiando a Terra Project"
should be the first author of the Zenodo archives and everyone else
should be in alphabetical order. Remove the step for checking author
order and included one to make sure the new guidelines are in place
(won't be needed anymore when all projects have converted).