Remove "Description" heading from PR template (#15)
When a PR is made with a single commit, the commit message is included at 
the top of the PR description. With our template, we had to copy that below
the "Description" heading, which was very annoying. This new template asks
for a description above and no heading, which eliminates that.