Update steps in release checklist
Update the steps in the release checklist after some modifications in
Zenodo UI. Also, instruct maintainers to write the changelog in Markdown
first, and then convert to RST: Markdown is easier to edit, less prone
to errors, and we already use Markdown for our commit messages.

These changes were inspired by the release process we took in São Paulo
in August 2024 with Arthur, Gelson, India, and Yago.